test(webhook): mirror apply_handlers split in test files
Rename apply_handlers_test.go to apply_gating_test.go and move
TestDDLMatchesStoredPlan into a new apply_execute_test.go so tests
sit beside the production files they exercise after the previous
split commit.

No test logic changes.

Validation:
- go test -count=1 -race ./pkg/webhook/...
- golangci-lint run --new-from-rev=origin/main ./pkg/webhook/... (0 issues)
